Add more exception types
... along with tests. Documentation in Exception.pm.
Note that while there are tests for the XS output, I have not tried
compiling it yet. I'm also not convinced that all of these exception
types are a good idea.
What's still missing is a simple way for the user to generate nice error
messages using the properties of the C++ exception object. The "code"
type can do this in raw XS, of course.